requirements.txt
# Core dependencies
anthropic>=0.18.0
litellm>=1.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
watchdog>=3.0.0

# Development dependencies
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0